Fixed wrong type initialization in VarUtils(I am sorry, my brain is not braining anymore)

+++ src/main/cpp/com/codingbee/utilities/VarUtils.cpp
@@ -29,7 +29,7 @@ void VarUtils::copyMatrix(int **matrix1, int **&matrix2, int rows, int columns){
 
 int VarUtils::getHighestIndexInSubVector(float *vector, int startingIndex, int endIndex){
     int highestIndex = startingIndex;
-    int highestValue = 0;
+    float highestValue = 0;
     for (int i = startingIndex + 1; i <= endIndex; i++){
         if (vector[i] > highestValue){
             highestIndex = i;
